maint.mk: add syntax-check to avoid char[PATH_MAX]
POSIX allows PATH_MAX to be undefined.  And even if you use the
gnulib pathmax module, where "pathmax.h" guarantees a definition,
the definition might not be constant or might be so large as to
be wasteful or cause stack overflows.  PATH_MAX should only be
used as a limit or hueristic, not an array size.

* top/maint.mk (sc_prohibit_path_max_array): New rule.

Signed-off-by: Eric Blake <eblake@redhat.com>