Add a new sethostname module
Define sethostname on platforms that do not provide the declaration.
Provide a function for platforms that lack it.  The general handling
of the provided function is to simply return -1 and set errno to
ENOSYS.  A specific handler is provided for Minix.

* lib/sethostname.c (sethostname): New file.  Provide sethostname
for systems that lack it.
* m4/sethostname.m4 (gl_FUNC_SETHOSTNAME): New file.  Detection of
sethostname declaration and function.
* modules/sethostname: New file.  Define the sethostname module.

Signed-off-by: Ben Walton <bwalton@artsci.utoronto.ca>