strptime: fix regression on mingw
The previous change lost the side effect of skipping past a
time-zone offset sign.  Instead, fix things to rely on the
fact that _GL_UNUSED can be applied to any variable, even if
some compilation paths use it, to silence gcc on the paths
where it is not used.

* lib/strptime.c (__strptime_internal) [!_LIBC && !HAVE_TM_GMTOFF]:
Fix regression.  Reported by Bruno Haible.

Signed-off-by: Eric Blake <eblake@redhat.com>