bootstrap: only force latest Makefile.in.in for gettext module
Commit 45fb479c48453d was added on behalf of bison, which uses the
'gettext' module, and where bison wanted to pick up the enhancements
to Makefile.in.in as provided by gnulib (such as an install-dvi
target).  This works because the 'gettext' module assumes that as
soon as upstream gettext moves to 0.19, you will also want to
update your development machine to the same version, so always
having the latest Makefile.in.in is reasonable.

But there are also projects like libvirt, that want to build out of
the box on active enterprise environments such as RHEL 5, which is
still at gettext 0.17.  The 'gettext-h' module is designed for these
projects, to pick up a useful "gettext.h" with _(), but to not force
the latest and greatest gettext installation.  However, the recent
bootstrap change broke these projects, since the versioning between
a project's configure.ac (0.17) and the gnulib version (0.18) cause
make to fail in the po directory.

The solution is to recognize both classes of projects - either you
care about latest GNU Coding Standards (and therefore use the
'gettext' module and the latest gettext release) or stability (and
therefore use the 'gettext-h' module and can get by with an older
release; although at this point in time, anything older than 0.17
doesn't seem to be in common enterprise use).

* build-aux/bootstrap (with_gettext): Only install latest
Makefile.in.in for projects requesting bleeding edge gettext.
Signed-off-by: Eric Blake <eblake@redhat.com>