maint.mk: let packages tweak sc_po_check pattern
libosinfo wants to look for XML files with translation markers
of the form <_name>string to translate</_name>.  Rather than
globally adding this to the regex, I thought it would be easier
to make it possible for cfg.mk to override the default pattern.
https://www.redhat.com/archives/virt-tools-list/2012-October/msg00270.html

* top/maint.mk (sc_po_check): Add translation-markers, to allow
finding files with other translation markers.

Signed-off-by: Eric Blake <eblake@redhat.com>