bootstrap: AC_INIT may have more than four parameters
For packages that did
  AC_INIT([package], [version], [bug-report], [], [url])
the extraction script was grabbing url instead of package for the
value of $package.

Furthermore, while we guarantee LC_ALL=C, this code is liable to
be copied and pasted to a script that doesn't use the C locale and
thus where a range expression may misbehave.

* build-aux/bootstrap (extract_package_name): Correctly extract
non-empty tarname field.  Avoid range in regex.
Based on a report by Sami Kerola <kerolasa@iki.fi>.

Signed-off-by: Eric Blake <eblake@redhat.com>